+++ build-packages/scandipwa-development-toolkit-vscode/src/commands/extend/index.ts
@@ -1,8 +1,8 @@
 import { extend, ResourceType } from "@scandipwa/scandipwa-development-toolkit-core";
-import { getSourceModule, getTargetModule } from "../options/module-selection";
+import { getSourceModule, getTargetModule } from "../../options/module-selection";
 // import { getResourceName } from "../options/resource-params";
-import logger from "../util/logger";
-import ui from '../util/ui';
+import logger from "../../util/logger";
+import ui from '../../util/ui';
 
 export const extender = (resourceType: ResourceType) => async () => {
     // resourceName,
